Move common types to superdesk.core.types
